Pass all shift tests

+++ gba-core/src/cpu/instrs/arm/branch_and_exchange.rs
@@ -14,7 +14,7 @@ impl ArmInstruction for BranchAndExchange {
             todo!("undefined behaviour");
         }
 
-        let dest = cpu.get_reg(rn as usize);
+        let dest = cpu.get_reg(rn);
 
         if dest.bit(0) == 1 {
             cpu.set_state(State::Thumb);
